requirements.txt
aiohttp==3.9.4
fastapi==0.110.1
GitPython==3.1.43
pydantic==2.6.4
